fixed paddle movement

+++ Pong/rectangle.cpp
@@ -1,7 +1,7 @@
+#include "olcPixelGameEngine.h"
+
 #include "shape.hpp"
 #include "rectangle.hpp"
-
-#include "olcPixelGameEngine.h"
 
 Rectangle::Rectangle(olc::PixelGameEngine* pge, olc::vf2d pos, olc::vf2d dir, int speed, int width, int height) : Shape (pge, pos, dir, speed) {
 	this->width = width;
@@ -9,6 +9,6 @@ Rectangle::Rectangle(olc::PixelGameEngine* pge, olc::vf2d pos, olc::vf2d dir, in
 }
 
 void Rectangle::draw() {
-	pge->DrawRect(pos.x, pos.y, width, height, { 255, 175, 181 });
-	pge->FillRect(pos.x, pos.y, width, height, { 255, 175, 181 });
+	pge->DrawRect((int)pos.x, (int)pos.y, width, height, { 255, 175, 181 });
+	pge->FillRect((int)pos.x, (int)pos.y, width, height, { 255, 175, 181 });
 }
